requirements.txt
# pip install -r requirements.txt
flask
python-dotenv
openai
apscheduler
markdown